Add scripts, lastpassify, database.example.yml
I was running into an error when starting the repo without a
`database.yml`. I needed `lastpassify` to create one.

While I was working on this I decided to make scripts to make setup and
shutdown easier, as after the time between projects I had forgotten how
exactly to do it.

Note: I'm not sure why our usual way of running `lastpassify` didn't
work. I left the usual way we have it in scripts as a comment because I
wanted to draw attention that it is not our usual way of doing this, and
perhaps the team would have feedback on why it's not working.

With all of these changes, now the project starts successfully.